started

+++ src/main/java/ru/netology/service/PostService.java
@@ -17,7 +17,7 @@ public class PostService {
     return repository.all();
   }
 
-  public Post getById(int id) {
+  public Post getById(long id) {
     return repository.getById(id).orElseThrow(NotFoundException::new);
   }
 
@@ -25,7 +25,7 @@ public class PostService {
     return repository.save(post);
   }
 
-  public void removeById(int id) {
+  public void removeById(long id) {
     repository.removeById(id);
   }
 }
